django/requirements.txt
Django==2.2.12
python-dotenv==0.13.0
pytz==2020.1
sqlparse==0.3.1
